+++ flash_examples/integrations/fiftyone/image_classification.py
@@ -37,7 +37,7 @@ datamodule = ImageClassificationData.from_folders(
 model = ImageClassifier(
     backbone="resnet18",
     num_classes=datamodule.num_classes,
-    serializer=Labels(),
+    output=Labels(),
 )
 trainer = flash.Trainer(
     max_epochs=1,
@@ -56,7 +56,7 @@ trainer.save_checkpoint("image_classification_model.pt")
 model = ImageClassifier.load_from_checkpoint(
     "https://flash-weights.s3.amazonaws.com/0.5.2/image_classification_model.pt"
 )
-model.serializer = FiftyOneLabels(return_filepath=True)  # output FiftyOne format
+model.output = FiftyOneLabels(return_filepath=True)  # output FiftyOne format
 predictions = trainer.predict(model, datamodule=datamodule)
 predictions = list(chain.from_iterable(predictions))  # flatten batches