+++ flash_examples/serve/tabular_classification/inference_server.py
@@ -17,5 +17,5 @@ from flash.tabular import TabularClassifier
 model = TabularClassifier.load_from_checkpoint(
     "https://flash-weights.s3.amazonaws.com/0.5.2/tabular_classification_model.pt"
 )
-model.serializer = Labels(["Did not survive", "Survived"])
+model.output = Labels(["Did not survive", "Survived"])
 model.serve()